+++ CHANGES.md
@@ -6,6 +6,7 @@ CHANGES
  * [base] add `local_html` for checking local fragment
  * [html] use `local_html` to check local fragment
  * add `print_toplist` for a list of worse post/entry/etc
+ * add HTML attribute `name` to search list of fragment and use regular expression to match
 
 ## Version 0.1.1 (2013-02-15T12:07:09Z)