+++ CHANGES.md
@@ -7,6 +7,7 @@ CHANGES
  * [html] use `local_html` to check local fragment
  * add `print_toplist` for a list of worse post/entry/etc
  * add HTML attribute `name` to search list of fragment and use regular expression to match
+ * fix #1: url not checked when same url with fragment also being checked
 
 ## Version 0.1.1 (2013-02-15T12:07:09Z)